Add Bugsnag params filtering
Sometimes we don't want to send all information to Bugsnag for security
reasons. For the values of the metadata to appear as [FILTERED], we need
to add rules detailing which properties are to be ignored.
This commit adds a couple of keys, shamelessly copied from MetricsAPI
project, as an example of how to apply the filters.