Disable Metrics/BlockLength cop in spec folder
The reason for this is that Rubocop now checks for more than
The Ruby Style guide, which helps enforcing conssitency in projects,
but doesn't know the diffence between plain Ruby code and DSL's.

In this case the issue is related to the way the RSpec DSL defines
specs and tests and not with a huge block of Ruby code.