Fix Gemfile.lock re-write
Due to first running the `ADD Gemfile* $APP_HOME/` we copy both Gemfile
and the Gemfile.lock.
After, the command `RUN bundle install` the Gemfile.lock is modified
with addition of the gem versions.
When the `COPY . $APP_HOME` command is ran, it re-writes the
Gemfile.lock with the previous version that did not contain the gem
versions, making the following command `bin/rspec`, for example, fail.