Add tests-jruby-debug docker-compose target using --debug flag
The --debug JRuby flag enables setting breakpoints and stepping when
using pry-debugger-jruby, but this option also has a negative impact on
performance and thus I've been recommending to other talkdeskers that
they don't use it in their tests target.

But as it is rather awkward to have to manually add it when you need to
debug, let's add a separate target that already includes it, so that
people aren't tempted to add it as a default.