Correct the production logging pattern date formatter
We noticed that no logging was available on the qa/prod environment of
one of our systems. The culprit appears to be the `date_method` used to
by the logger to format the date.

Internally the [logging library was freezing the time instance][1] it creates for the log event
and the `#utc` invocation tries to mutate the datastructure (which
raises an exception and silences the app logging).

As an alternative, make use of the [utc_offset][2] option to make sure all logs are stored in the UTC timezone.

[1]:
https://github.com/TwP/logging/blob/f9150dd134c1b615832c5109e14494782937dfa3/lib/logging/layouts/pattern.rb#L221
[2]:
https://github.com/TwP/logging/blob/f9150dd134c1b615832c5109e14494782937dfa3/lib/logging/layout.rb#L80